Rewrite multi-touch implementation (breaking!)
The old way made it pretty hard to track touches
in my opinion. There was no way of detecting a
released finger in a clean way.

This implementation kept most of the old structure
and changed one "Touch" event to three:
"Press", "Release" and "Move".
I reused my implementation I've done for a port
of plato to the reMarkable from here:
https://github.com/LinusCDE/plato/blob/4d9c3444/src/input.rs#L364

That particular implementation was already
tested to work good with multiple fingers at
once. The same should apply to this commit.

The previous implementation was probably more
performant, but I don't consider it paramount.
This change was able to produce some 0.7%
spiked with a lot of multitouch-event
where the original implementation didn't
increase at all.